feat: directions from one class to another | refactor algorithms, add dijkstra test cases (#40)
