+++ components/IndoorDirections/TypesOfDirections/DifferentBuildingDirections.js
@@ -1,12 +1,25 @@
 import React from 'react';
 import { Line, G } from 'react-native-svg';
 import { ClassGraph } from '../../../constants/ClassGraph';
+import { dijkstra, getFloorNumber, changeClassName } from '../Dijkstra/DijkstraAlgorithm';
 
 export function DifferentBuildingDirections(props) {
     const rooms = props.rooms;
     const graph = ClassGraph();
+    var goTo = props.from;
+    var tempGoTo = props.from;
 
-    const pathToElevator = dijkstra(graph, props.from, "elevator").path;
+    if (props.from.includes(" ")) {
+        tempGoTo = goTo;
+        goTo = changeClassName(props.to);
+    }
+
+    if (props.to.includes(" ")) {
+        tempGoTo = goTo;
+        goTo = changeClassName(props.from);
+    }
+
+    const pathToElevator = dijkstra(graph, goTo, "elevator").path;
     const pathToClass = dijkstra(graph, "elevator", "exit").path;
 
     var getNextRoomElevator = (index) => {
@@ -37,10 +50,10 @@ export function DifferentBuildingDirections(props) {
     
 
     // Go to elevator
-    if(props.floor == getFloorNumber(props.from)) {
+    if(props.floor == getFloorNumber(tempGoTo)) {
         return(
             <G>
-                <Line x1={rooms[props.from].x} y1={rooms[props.from].y} x2={rooms[props.from].nearestPoint.x} y2={rooms[props.from].nearestPoint.y} stroke="blue" strokeWidth="5"/>
+                <Line x1={rooms[goTo].x} y1={rooms[goTo].y} x2={rooms[goTo].nearestPoint.x} y2={rooms[goTo].nearestPoint.y} stroke="blue" strokeWidth="5"/>
                 {
                     linesToElevator.map(el => <Line x1={el.x1} y1={el.y1} x2={el.x2} y2={el.y2} stroke="blue" strokeWidth="5"/>)
                 }
@@ -62,82 +75,3 @@ export function DifferentBuildingDirections(props) {
         <G></G>
     )
 }
-
-/**
- * Get the floor number of a specific classroom. 
- * Ex: H829 returns 8
- * @param {*} name | classroom name
- */
-export const getFloorNumber = (name) => {
-    if (name) {
-        const num = name.match(/\d+/g).map(Number);
-
-        if(num.length > 1) {
-            return num[0].toString();
-        }
-        if (num.toString().length == 3) {
-            return num.toString().charAt(0)
-        }
-        if (num.toString().length == 4) {
-            return num.toString().charAt(0) + num.toString().charAt(1);
-        }   
-    }
-}
-
-const dijkstra = (graph, start, end) => {
-    // track the lowest cost to reach each node
-    let costs = {};
-    costs[end] = "Infinity";
-    costs = Object.assign(costs, graph[start]);
-
-    // track paths
-    const parents = {end: null};
-    for (let child in graph[start]) {
-        parents[child] = start;
-    }
-
-    const processed = [];
-    let node = lowestCostNode(costs, processed);
-
-    while (node) {
-        let cost = costs[node];
-        let children = graph[node];
-        for (let n in children) {
-            if (String(n) !== String(start)) {
-                let newCost = cost + children[n];
-                if (!costs[n] || costs[n] > newCost) {
-                    costs[n] = newCost;
-                    parents[n] = node;
-                }
-            }
-        }
-        processed.push(node);
-        node = lowestCostNode(costs, processed);
-    }
-
-    let optimalPath = [end];
-    let parent = parents[end];
-    while (parent) {
-        optimalPath.push(parent);
-        parent = parents[parent];
-    }
-    optimalPath.reverse();
-
-    const results = {
-        distance: costs[end],
-        path: optimalPath
-    };
-
-    return results;
-};
-
-const lowestCostNode = (costs, processed) => {
-    return Object.keys(costs).reduce((lowest, node) => {
-        if (lowest === null || costs[node] < costs[lowest]) {
-            if (!processed.includes(node)) {
-                lowest = node;
-            }
-        }
-        return lowest;
-    }, null);
-};
